+++ integrations/authorization/v1/configuration.go
@@ -70,14 +70,14 @@ func (c Configuration) Plugin(ctx context.Context) (pbImplAuthorizationV1Shared.
 		if err != nil {
 			return nil, err
 		}
-		return sidecarSvcAuthzClient.NewClient(ctx, sidecarSvcAuthzDefinition.NewAuthorizationPoolServiceClient(conn)), nil
+		return pbImplAuthorizationV1Shared.SuperUser(sidecarSvcAuthzClient.NewClient(ctx, sidecarSvcAuthzDefinition.NewAuthorizationPoolServiceClient(conn))), nil
 
 	case ConfigurationTypeCentralPermissive:
 		conn, err := integration.NewIntegrationConnection()
 		if err != nil {
 			return nil, err
 		}
-		return pbImplAuthorizationV1Shared.Permissive(sidecarSvcAuthzClient.NewClient(ctx, sidecarSvcAuthzDefinition.NewAuthorizationPoolServiceClient(conn)), logger), nil
+		return pbImplAuthorizationV1Shared.SuperUser(pbImplAuthorizationV1Shared.Permissive(sidecarSvcAuthzClient.NewClient(ctx, sidecarSvcAuthzDefinition.NewAuthorizationPoolServiceClient(conn)), logger)), nil
 	default:
 		return nil, errors.Errorf("Configuration type '%s' is not supported", string(c.Type))
 	}