+++ pkg/sidecar/services/authorization/impl_api_roles.go
@@ -32,13 +32,12 @@ import (
 )
 
 func (a *implementation) APIGetRole(ctx context.Context, request *sidecarSvcAuthzDefinition.AuthorizationAPINamedRequest) (*sidecarSvcAuthzDefinition.AuthorizationAPIRoleResponse, error) {
-	if identity := authenticator.GetIdentity(ctx); identity == nil {
-		return nil, status.Error(codes.Unauthenticated, "Unauthenticated")
-	} else {
-		// Restrict only for superuser for now
-		if identity.User != nil {
-			return nil, status.Error(codes.Unauthenticated, "Only super-user allowed")
-		}
+	if err := a.Health(ctx).Require(); err != nil {
+		return nil, err
+	}
+
+	if err := authenticator.GetIdentity(ctx).EvaluatePermission(ctx, a.auth, "rbac:GetRole", request.GetName()); err != nil {
+		return nil, err
 	}
 
 	role, index, ok := a.roles.Item(request.GetName())
@@ -54,13 +53,12 @@ func (a *implementation) APIGetRole(ctx context.Context, request *sidecarSvcAuth
 }
 
 func (a *implementation) APIDeleteRole(ctx context.Context, request *sidecarSvcAuthzDefinition.AuthorizationAPINamedRequest) (*sidecarSvcAuthzDefinition.AuthorizationAPIRoleResponse, error) {
-	if identity := authenticator.GetIdentity(ctx); identity == nil {
-		return nil, status.Error(codes.Unauthenticated, "Unauthenticated")
-	} else {
-		// Restrict only for superuser for now
-		if identity.User != nil {
-			return nil, status.Error(codes.Unauthenticated, "Only super-user allowed")
-		}
+	if err := a.Health(ctx).Require(); err != nil {
+		return nil, err
+	}
+
+	if err := authenticator.GetIdentity(ctx).EvaluatePermission(ctx, a.auth, "rbac:DeleteRole", request.GetName()); err != nil {
+		return nil, err
 	}
 
 	offset, err := a.roles.Delete(ctx, request.GetName())
@@ -80,13 +78,12 @@ func (a *implementation) APIDeleteRole(ctx context.Context, request *sidecarSvcA
 }
 
 func (a *implementation) APICreateRole(ctx context.Context, request *sidecarSvcAuthzDefinition.AuthorizationAPIRoleRequest) (*sidecarSvcAuthzDefinition.AuthorizationAPIRoleResponse, error) {
-	if identity := authenticator.GetIdentity(ctx); identity == nil {
-		return nil, status.Error(codes.Unauthenticated, "Unauthenticated")
-	} else {
-		// Restrict only for superuser for now
-		if identity.User != nil {
-			return nil, status.Error(codes.Unauthenticated, "Only super-user allowed")
-		}
+	if err := a.Health(ctx).Require(); err != nil {
+		return nil, err
+	}
+
+	if err := authenticator.GetIdentity(ctx).EvaluatePermission(ctx, a.auth, "rbac:CreateRole", request.GetName()); err != nil {
+		return nil, err
 	}
 
 	if item := request.GetItem(); item == nil {
@@ -111,13 +108,12 @@ func (a *implementation) APICreateRole(ctx context.Context, request *sidecarSvcA
 }
 
 func (a *implementation) APIUpdateRole(ctx context.Context, request *sidecarSvcAuthzDefinition.AuthorizationAPIRoleRequest) (*sidecarSvcAuthzDefinition.AuthorizationAPIRoleResponse, error) {
-	if identity := authenticator.GetIdentity(ctx); identity == nil {
-		return nil, status.Error(codes.Unauthenticated, "Unauthenticated")
-	} else {
-		// Restrict only for superuser for now
-		if identity.User != nil {
-			return nil, status.Error(codes.Unauthenticated, "Only super-user allowed")
-		}
+	if err := a.Health(ctx).Require(); err != nil {
+		return nil, err
+	}
+
+	if err := authenticator.GetIdentity(ctx).EvaluatePermission(ctx, a.auth, "rbac:UpdateRole", request.GetName()); err != nil {
+		return nil, err
 	}
 
 	if item := request.GetItem(); item == nil {