+++ pkg/util/cache/object.go
@@ -39,6 +39,7 @@ type ObjectFetcher[T any] func(ctx context.Context) (T, time.Duration, error)
 type Object[T any] interface {
 	Init(context.Context) error
 	Get(ctx context.Context) (T, error)
+	GetWithTTL(ctx context.Context) (T, time.Duration, error)
 }
 
 type object[T any] struct {
@@ -54,24 +55,28 @@ func (o *object[T]) Init(ctx context.Context) error {
 	_, err := o.Get(ctx)
 	return err
 }
-
 func (o *object[T]) Get(ctx context.Context) (T, error) {
+	v, _, err := o.GetWithTTL(ctx)
+	return v, err
+}
+
+func (o *object[T]) GetWithTTL(ctx context.Context) (T, time.Duration, error) {
 	o.lock.Lock()
 	defer o.lock.Unlock()
 
 	if time.Now().After(o.eol) || o.eol.IsZero() {
 		obj, ttl, err := o.caller(ctx)
 		if err != nil {
-			return util.Default[T](), err
+			return util.Default[T](), 0, err
 		}
 
 		if ttl <= 0 {
-			return obj, nil
+			return obj, 0, nil
 		}
 
 		o.obj = obj
 		o.eol = time.Now().Add(ttl)
 	}
 
-	return o.obj, nil
+	return o.obj, time.Until(o.eol), nil
 }