+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - (Feature) Introduce ArangoD Data Helper
 - (Maintenance) Bump Kubernetes to 1.33
 - (Feature) (Platform) Authorization Central Service
+- (Feature) (Platform) Permission RBAC
 
 ## [1.4.1](https://github.com/arangodb/kube-arangodb/tree/1.4.1) (2026-01-27)
 - (Bugfix) (Platform) Ensure Helm History limit in ArangoPlatformService